requirements.txt
jira==3.8.0
python-dotenv==1.1.0
requests==2.32.3
PyGithub==2.6.1
GitPython==3.1.44
urllib3==2.4.0
six==1.17.0
aider-chat==0.83.1
PyYAML==6.0.2
ftfy==6.3.1